+++ src/components/builder.ts
@@ -15,16 +15,16 @@ const texMagicProgramName = 'TeXMagicProgram'
 const bibMagicProgramName = 'BibMagicProgram'
 
 export class Builder {
-    extension: Extension
-    tmpDir: string
-    currentProcess: cp.ChildProcessWithoutNullStreams | undefined
+    private readonly extension: Extension
+    readonly tmpDir: string
+    private currentProcess: cp.ChildProcessWithoutNullStreams | undefined
     disableBuildAfterSave: boolean = false
     disableCleanAndRetry: boolean = false
-    buildMutex: Mutex
-    waitingForBuildToFinishMutex: Mutex
-    isMiktex: boolean = false
-    previouslyUsedRecipe: {name: string, tools: (string | StepCommand)[]} | undefined
-    previousLanguageId: string | undefined
+    private readonly buildMutex: Mutex
+    private readonly waitingForBuildToFinishMutex: Mutex
+    private readonly isMiktex: boolean = false
+    private previouslyUsedRecipe: Recipe | undefined
+    private previousLanguageId: string | undefined
 
     constructor(extension: Extension) {
         this.extension = extension
@@ -47,6 +47,9 @@ export class Builder {
         }
     }
 
+    /**
+     * Kills the current building process.
+     */
     kill() {
         const proc = this.currentProcess
         if (proc) {
@@ -70,11 +73,11 @@ export class Builder {
         return this.buildMutex.count === 1
     }
 
-    isWaitingForBuildToFinish(): boolean {
+    private isWaitingForBuildToFinish(): boolean {
         return this.waitingForBuildToFinishMutex.count < 1
     }
 
-    async preprocess(): Promise<() => void> {
+    private async preprocess(): Promise<() => void> {
         const configuration = vscode.workspace.getConfiguration('latex-workshop')
         this.disableBuildAfterSave = true
         await vscode.workspace.saveAll()
@@ -85,6 +88,14 @@ export class Builder {
         return releaseBuildMutex
     }
 
+    /**
+     * Executes a command building LaTeX files.
+     *
+     * @param command The name of the command to build LaTeX files.
+     * @param args The arguments of the command.
+     * @param pwd The path of the working directory of building.
+     * @param rootFile The root file to be compiled.
+     */
     async buildWithExternalCommand(command: string, args: string[], pwd: string, rootFile: string | undefined = undefined) {
         if (this.isWaitingForBuildToFinish()) {
             return
@@ -161,16 +172,23 @@ export class Builder {
         })
     }
 
-    buildInitiator(rootFile: string, languageId: string, recipe: string | undefined = undefined, releaseBuildMutex: () => void) {
-        const steps = this.createSteps(rootFile, languageId, recipe)
+    private buildInitiator(rootFile: string, languageId: string, recipeName: string | undefined = undefined, releaseBuildMutex: () => void) {
+        const steps = this.createSteps(rootFile, languageId, recipeName)
         if (steps === undefined) {
             this.extension.logger.addLogMessage('Invalid toolchain.')
             return
         }
-        this.buildStep(rootFile, steps, 0, recipe || 'Build', releaseBuildMutex) // use 'Build' as default name
+        this.buildStep(rootFile, steps, 0, recipeName || 'Build', releaseBuildMutex) // use 'Build' as default name
     }
 
-    async build(rootFile: string, languageId: string, recipe: string | undefined = undefined) {
+    /**
+     * Builds a LaTeX file with user-defined recipes.
+     *
+     * @param rootFile The root file to be compiled.
+     * @param languageId The name of the language of a file to be compiled.
+     * @param recipeName The name of a recipe to be used.
+     */
+    async build(rootFile: string, languageId: string, recipeName: string | undefined = undefined) {
         if (this.isWaitingForBuildToFinish()) {
             this.extension.logger.addLogMessage('Another LaTeX build processing is already waiting for the current LaTeX build to finish. Exit.')
             return
@@ -216,7 +234,7 @@ export class Builder {
                 const relativePath = path.dirname(file.replace(rootDir, '.'))
                 fs.ensureDirSync(path.resolve(outDir, relativePath))
             })
-            this.buildInitiator(rootFile, languageId, recipe, releaseBuildMutex)
+            this.buildInitiator(rootFile, languageId, recipeName, releaseBuildMutex)
         } catch (e) {
             this.extension.logger.addLogMessage('Unexpected Error: please see the console log of the Developer Tools of VS Code.')
             this.extension.logger.displayStatus('x', 'errorForeground')
@@ -226,7 +244,7 @@ export class Builder {
         }
     }
 
-    progressString(recipeName: string, steps: StepCommand[], index: number) {
+    private progressString(recipeName: string, steps: StepCommand[], index: number) {
         if (steps.length < 2) {
             return recipeName
         } else {
@@ -234,7 +252,7 @@ export class Builder {
         }
     }
 
-    buildStep(rootFile: string, steps: StepCommand[], index: number, recipeName: string, releaseBuildMutex: () => void) {
+    private buildStep(rootFile: string, steps: StepCommand[], index: number, recipeName: string, releaseBuildMutex: () => void) {
         if (index === 0) {
             this.extension.logger.clearCompilerMessage()
         }
@@ -360,7 +378,7 @@ export class Builder {
         })
     }
 
-    buildFinished(rootFile: string) {
+    private buildFinished(rootFile: string) {
         this.extension.buildInfo.buildEnded()
         this.extension.logger.addLogMessage(`Successfully built ${rootFile}.`)
         this.extension.logger.displayStatus('check', 'statusBar.foreground', 'Recipe succeeded.')
@@ -382,7 +400,7 @@ export class Builder {
         }
     }
 
-    createSteps(rootFile: string, languageId: string, recipeName: string | undefined): StepCommand[] | undefined {
+    private createSteps(rootFile: string, languageId: string, recipeName: string | undefined): StepCommand[] | undefined {
         let steps: StepCommand[] = []
         const configuration = vscode.workspace.getConfiguration('latex-workshop')
 
@@ -402,22 +420,20 @@ export class Builder {
                 steps = [magicTex]
             }
         } else {
-            const recipes = configuration.get('latex.recipes') as {name: string, tools: (string | StepCommand)[]}[]
+            const recipes = configuration.get('latex.recipes') as Recipe[]
+            const defaultRecipeName = configuration.get('latex.recipe.default') as string
             const tools = configuration.get('latex.tools') as StepCommand[]
             if (recipes.length < 1) {
                 this.extension.logger.addLogMessage('No recipes defined.')
                 this.extension.logger.showErrorMessage('No recipes defined.')
                 return undefined
             }
-            let recipe = recipes[0]
+            let recipe: Recipe | undefined = undefined
             if (this.previousLanguageId !== languageId) {
                 this.previouslyUsedRecipe = undefined
             }
-            const defaultRecipe = configuration.get('latex.recipe.default') as string
-            if ((defaultRecipe === 'lastUsed') && (this.previouslyUsedRecipe !== undefined)) {
-                recipe = this.previouslyUsedRecipe
-            } else if ((defaultRecipe !== 'first') && (defaultRecipe !== 'lastUsed') && recipeName === undefined) {
-               recipeName = defaultRecipe
+            if (!recipeName && ! ['first', 'lastUsed'].includes(defaultRecipeName)) {
+                recipeName = defaultRecipeName
             }
             if (recipeName) {
                 const candidates = recipes.filter(candidate => candidate.name === recipeName)
@@ -427,13 +443,26 @@ export class Builder {
                 }
                 recipe = candidates[0]
             }
-            if (!recipeName && languageId === 'rsweave') {
-                const candidates = recipes.filter(candidate => candidate.name.toLowerCase().match('rnw|rsweave'))
-                if (candidates.length < 1) {
-                    this.extension.logger.addLogMessage(`Failed to resolve build recipe: ${recipeName}`)
-                    this.extension.logger.showErrorMessage(`Failed to resolve build recipe: ${recipeName}`)
+            if (recipe === undefined) {
+                if (defaultRecipeName === 'lastUsed') {
+                    recipe = this.previouslyUsedRecipe
                 }
-                recipe = candidates[0]
+                if (defaultRecipeName === 'first' || recipe === undefined) {
+                   let candidates: Recipe[] = recipes
+                   if (languageId === 'rsweave') {
+                        candidates = recipes.filter(candidate => candidate.name.toLowerCase().match('rnw|rsweave'))
+                   } else if (languageId === 'jlweave') {
+                        candidates = recipes.filter(candidate => candidate.name.toLowerCase().match('jnw|jlweave|weave.jl'))
+                   }
+                    if (candidates.length < 1) {
+                        this.extension.logger.addLogMessage(`Failed to resolve build recipe: ${recipeName}`)
+                        this.extension.logger.showErrorMessage(`Failed to resolve build recipe: ${recipeName}`)
+                    }
+                    recipe = candidates[0]
+                }
+            }
+            if (recipe === undefined) {
+                return undefined
             }
             this.previouslyUsedRecipe = recipe
             this.previousLanguageId = languageId
@@ -442,7 +471,7 @@ export class Builder {
                 if (typeof tool === 'string') {
                     const candidates = tools.filter(candidate => candidate.name === tool)
                     if (candidates.length < 1) {
-                        this.extension.logger.showErrorMessage(`Skipping undefined tool "${tool}" in recipe "${recipe.name}."`)
+                        this.extension.logger.showErrorMessage(`Skipping undefined tool "${tool}" in recipe "${recipe?.name}."`)
                     } else {
                         steps.push(candidates[0])
                     }
@@ -495,7 +524,7 @@ export class Builder {
         return steps
     }
 
-    findProgramMagic(rootFile: string): [StepCommand | undefined, StepCommand | undefined] {
+    private findProgramMagic(rootFile: string): [StepCommand | undefined, StepCommand | undefined] {
         const regexTex = /^(?:%\s*!\s*T[Ee]X\s(?:TS-)?program\s*=\s*([^\s]*)$)/m
         const regexBib = /^(?:%\s*!\s*BIB\s(?:TS-)?program\s*=\s*([^\s]*)$)/m
         const regexTexOptions = /^(?:%\s*!\s*T[Ee]X\s(?:TS-)?options\s*=\s*(.*)$)/m
@@ -546,4 +575,9 @@ interface StepCommand {
     command: string,
     args?: string[],
     env?: ProcessEnv
+}
+
+interface Recipe {
+    name: string,
+    tools: (string | StepCommand)[]
 }