+++ src/providers/preview/graphicsscaler.ts
@@ -1,16 +1,13 @@
-import {Extension} from '../../main'
 import * as path from 'path'
 import * as workerpool from 'workerpool'
 import {Proxy} from 'workerpool'
 import {IGraphicsScalerWorker} from './graphicsscaler_worker'
 
 export class GraphicsScaler {
-    extension: Extension
-    pool: workerpool.WorkerPool
-    proxy: workerpool.Promise<Proxy<IGraphicsScalerWorker>>
+    private readonly pool: workerpool.WorkerPool
+    private readonly proxy: workerpool.Promise<Proxy<IGraphicsScalerWorker>>
 
-    constructor(extension: Extension) {
-        this.extension = extension
+    constructor() {
         this.pool = workerpool.pool(
             path.join(__dirname, 'graphicsscaler_worker.js'),
             { maxWorkers: 1, workerType: 'process' }