+++ test/viewer.test.ts
@@ -182,7 +182,7 @@ suite('PDF Viewer test suite', () => {
         for (const result of secondResults) {
             assert.ok( Math.abs(result.scrollTop) > 10 )
         }
-    }, () => os.platform() !== 'linux')
+    }, () => os.platform() === 'win32')
 
     runTestWithFixture('fixture021', 'basic build, view, and synctex with synctex.afterBuild.enabled', async () => {
         const fixtureDir = getFixtureDir()
@@ -217,5 +217,5 @@ suite('PDF Viewer test suite', () => {
         for (const result of secondResults) {
             assert.ok( Math.abs(result.scrollTop) > 10 )
         }
-    }, () => os.platform() !== 'linux')
+    }, () => os.platform() === 'win32')
 })