+++ CHANGELOG.md
@@ -1,5 +1,32 @@
 # Change Log
 
+## [8.11.1] - 2020-07-01
+
+### Fixed
+- Restoring PDF viewers does not work well with VS Code 1.47.
+- (#2172) `"latex-workshop.latex.recipe.default": "lastUsed"` doesn't work.
+- (#2176) No need to put the closing `}` on its own line in `.bib`.
+- (#2177) Use `kpsewhich` to resolve bibliography files.
+- (#2182) Also search for `.bib` files in `rootDir`.
+
+## [8.11.0] - 2020-06-24
+
+### Added
+- (#2149) Add support for `weave.jl`.
+- (#2154) Add section numbers in outline.
+  - This feature can be deactivated by setting `latex-workshop.view.outline.numbers.enabled: false`.
+
+### Improved
+- (#2109) Improve PDF viewer refresh speed.
+- (#2157) Improve the invert mode of the PDF viewer with additional filters.
+
+### Fixed
+- Activate `onEnterKey` even when suggestion is visible.
+  - Always activate `onEnterKey` when `acceptSuggestionOnEnter`.
+  - (#2167) Do not activate `onEnterKey` when `suggestWidgetVisible`.
+- (#2107) Use `fs.realpathSync` to compare file paths for SyncTeX.
+- (#2146) Use `rootFile` to expand placeholders when formatting.
+
 ## [8.10.0] - 2020-06-06
 
 ### Added